[Profiler] Extend ID assignment to allocations and frees
This is necessary for memory profiling because we need to know how to interpret an allocation. However there is a slight wrinkle: we don't know if an allocation is for a Tensor's StorageImpl until we see it used in a later call. (We could record outputs, however we're not willing to incur the overhead.) So we instead treat all allocations as relevant and then filter out some later. Otherwise the change to the ID assignment algorithm is minimal.

Differential Revision: [D39788870](https://our.internmc.facebook.com/intern/diff/D39788870/)

ghstack-source-id: 168584763
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85719